ref: Rename onboardingCard.tsx to match the name of the default export
